+++ Domain/Entities/ProductReview.cs
@@ -8,7 +8,7 @@ namespace Domain.Entities
 {
     public class ProductReview
     {
-        public ProductReview(double rating, string comment, string userId, string userEmail)
+        public ProductReview(int rating, string comment, string userId, string userEmail)
         {
             Rating = rating;
             Comment = comment;
@@ -17,7 +17,7 @@ namespace Domain.Entities
         }
         public int Id { get; set; }
         public string? Comment { get; set; }
-        public double Rating { get; set; }
+        public int Rating { get; set; }
         public int ProductId { get; set; }
         public Product Product { get; set; }
         public string? UserId { get; set; }
